refactor: TestUtil 도메인 단위로 분리
