+++ src/test/java/nextstep/subway/acceptance/StationAcceptanceTest.java
@@ -1,17 +1,13 @@
 package nextstep.subway.acceptance;
 
-import static nextstep.subway.acceptance.SubwayTestUtils.createStationWithName;
+import static nextstep.subway.acceptance.util.StationTestUtils.createStationWithName;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import io.restassured.RestAssured;
-import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
-import java.util.HashMap;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
 @DisplayName("지하철역 관련 기능")
 class StationAcceptanceTest extends AcceptanceTest {
