+++ src/test/java/nextstep/subway/acceptance/util/LineTestUtils.java
@@ -1,28 +1,13 @@
-package nextstep.subway.acceptance;
+package nextstep.subway.acceptance.util;
 
 import io.restassured.RestAssured;
 import io.restassured.mapper.ObjectMapperType;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import java.util.HashMap;
 import nextstep.subway.applicaion.dto.LineCreationRequest;
 import org.springframework.http.MediaType;
 
-public class SubwayTestUtils {
-
-    public static ExtractableResponse<Response> createStationWithName(String stationName) {
-        var requestBody = new HashMap<String, String>();
-        requestBody.put("name", stationName);
-
-        return RestAssured
-                .given()
-                    .body(requestBody)
-                    .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                    .post("/stations")
-                .then()
-                    .extract();
-    }
+public class LineTestUtils {
 
     public static ExtractableResponse<Response> createLine(LineCreationRequest creationRequest) {
         return RestAssured
